+++ frontend/src/components/CategoryFilterMenu/CategoryFilterMenu.style.ts
