+++ frontend/src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as Button } from './Button/Button';
-export { default as CategoryMenu } from './CategoryMenu/CategoryMenu';
+export { default as CategoryFilterMenu } from './CategoryFilterMenu/CategoryFilterMenu';
 export { default as Dropdown } from './Dropdown/Dropdown';
 export { default as Flex } from './Flex/Flex';
 export { default as Header } from './Header/Header';
